virida_reporting_api_service/requirements.txt
fastapi==0.61.1
uvicorn~=0.13.4
pydantic~=1.8
sqlalchemy~=1.3.23
pymysql~=0.10.0
cryptography~=3.0
pyjwt~=2.0.1
alembic~=1.4.2
requests~=2.24.0
aiohttp~=3.7.2
pathvalidate~=2.4.1
pandas~=1.1.4
sendgrid~=6.6.0
